# HannaCodeServer/docs/swagger.yaml
openapi: 3.0.0
info:
  title: HannaCode API
  version: 1.0.0
  description: API documentation for HannaCode platform
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
paths:
...
  /lessons:
    get:
      summary: Get all lessons
      tags: [Lessons]
      parameters:
        - in: query
          name: course
          schema:
            type: string
          description: Course ID
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Items per page
      responses:
        '200':
          description: List of lessons

    post:
      summary: Create a new lesson
      tags: [Lessons]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - content
                - course
                - order
              properties:
                title:
                  type: string
                description:
                  type: string
                content:
                  type: string
                course:
                  type: string
                order:
                  type: integer
      responses:
        '201':
          description: Lesson created successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...